verify the client user matches the session/cookie
add `username` property to `DimClient` to be able to verify the user
against DIM, this is to detect mismatches that may occur, when using
the same cookie file with multiple users.

Also set `HOME` for `make tests`, so it puts that in the `VDIR`
to allow us clearing the cookies before tests.
Not clearing the cookies results in DIM not automatically creating the
users, as it was only checked whether the cookie was valid, but that
the user is present in the DIM database.

ndcli currently allows us to set the cookie file, but not a cookie path,
the tests also assume they can change user by just doing `ndcli login`,
which requires to use *one* cookie file